requirements.txt
Django==3.1.6
pillow
flake8
django-allauth
django-ckeditor
django-widget_tweaks
xlwt
django-rest_framework
python-decouple
